Replace mocked FactoryGirl with "real" FactoryGirl
The problems with stubbing FactoryGirl are:

1. This feature spec should be testing the full stack rather than
mocking/stubbing out interfaces;
2. The mocked FactoryGirl’s .create method created a user resource
using User.create(opts). The problem is that Rails 4 requires that
attributes be whitelisted for mass-assignment with strong parameters. I
didn’t want to implement the Rails 4 strong parameters because it is
possible that some users of this gem are using a version of Rails that
doesn’t support strong parameters. The ‘real’ FactoryGirl avoids this
problem