Parse response body
Now that we are creating the user resource with factory_girl, the
`response.body` is returning more than just `{“first_name”:”Sam”}`.
Instead it is returning json like:
`"{\"id\":1,\"first_name\":\"Sam\",\"created_at\":\"2014-11-13T21:42:16.
205Z\",\"updated_at\":\"2014-11-13T21:42:16.205Z\”}”`

Instead of comparing the json as string equality (which is difficult
because of the changing date attributes), we are parsing the JSON and
comparing key/values.